Fix for wrong datataypes

+++ src/main/java/com/gmail/nossr50/commands/skills/HerbalismCommand.java
@@ -46,10 +46,10 @@ public class HerbalismCommand extends SkillCommand {
         greenTerraLength = String.valueOf(2 + (int) ((double) skillValue / (double) abilityLengthIncreaseLevel));
         //FARMERS DIET
         if(skillValue >= farmersDietMaxLevel) farmersDietRank = "5";
-        else farmersDietRank = String.valueOf((double) skillValue / (double) farmersDietRankChange);
+        else farmersDietRank = String.valueOf((int) ((double) skillValue / (double) farmersDietRankChange));
         //GREEN THUMB
         if(skillValue >= greenThumbStageMaxLevel) greenThumbStage = "4";
-        else greenThumbStage = String.valueOf((double) skillValue / (double) greenThumbStageChange);
+        else greenThumbStage = String.valueOf((int) ((double) skillValue / (double) greenThumbStageChange));
 
 
         if(skillValue >= greenThumbMaxLevel) greenThumbChance = String.valueOf(greenThumbMaxBonus);
